requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-filter==1.0.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
gunicorn==19.9.0
Pillow==3.1.2
whitenoise==4.1
Markdown==3.0.1